pyproject.toml
[project]
name = "it2ag"
version = "0.1.0"
description = "iTerm2 agent monitor for Claude Code and Codex"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "iterm2>=2.7,<3",
    "aiohttp>=3.9,<3.13",
    "gitpython>=3.1.46",
]

[project.scripts]
it2ag = "it2ag.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/it2ag"]

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.14",
    "pytest>=8",
    "pytest-asyncio>=0.25",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "B",
    "SIM",
    "RUF",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["iterm2", "iterm2.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
